refactor: Phase 2 - global-init.ts 5分割
1350行の global-init.ts を責務ベースで5ファイルに分割:
- global-config.ts (197行): 設定YAML読み書き、環境状態管理
- global-deploy.ts (353行): テンプレートコピー、maidctlデプロイ、インストール関数
- global-steps.ts (205行): buildExecutionSteps() + ステップ定義
- pending-state.ts (72行): VS Code再起動跨ぎのpending state管理
- global-init.ts (600行): 4フェーズ制御フロー（エントリポイント）

ロジック変更なし。global-init.ts からの re-export で後方互換を維持。
